// src/app/main/reports/report-type.ts
export type DurationType = 'month' | 'week' | 'day';
// ...
export type FilterState = {
  startDate: moment.Moment;
  endDate: moment.Moment;
  duration?: DurationType;
  group1?: string;
  group2?: string;
  machines?: number[];
  downtimeCodes?: number[];
  delayCodes?: number[];
  shifts?: number[];
  scheduleStatus?: string;
  isBack?: boolean;
};
// ...
export type DowntimeCodeDef = {
  code: number;
  category: 'None' | 'Change' | 'Breakdown' | 'OtherExpected' | 'OtherUnexpected' | 'Unscheduled';
  description: string;
};